Fix HOME resolution broken on Linux
On Linux systems we do not use HOME as a placeholder, it's either a
variable ($HOME) or using tilde (~ or ~user).

Moreover HOME was matched case insensitive, leading
to path like /home/myuser/myvsmods/ to be converted to
/home/myuser/mysuer/myvamods/ and then fails (silently).

Using expanduser() and expandvars() we can achieve the same results
without bugs on both Linux and Windows.